+++ src/playlistparser/__init__.py
@@ -10,9 +10,18 @@ from .track import Track
 
 
 class PlaylistParser(object):
-    def __init__(self, file_path=None, *, file_obj=None, require_fp=False, verbose=False):
+    def __init__(
+        self,
+        file_path=None,
+        *,
+        require_title=True,
+        require_duration=False,
+        require_year=False,
+        require_bpm=False,
+        require_fp=False,
+        verbose=False,
+    ):
         self.file_path = file_path
-        self.file_obj = file_obj
         self.file_contents = None
         self.verbose = verbose
         self.playlist_type = "none"
@@ -20,6 +29,10 @@ class PlaylistParser(object):
         self.tracks: List[Track] = []
         self._parser: Union[Callable[[str], List], None] = None
         self.is_parsed = False
+        self.require_title = require_title
+        self.require_duration = require_duration
+        self.require_bpm = require_bpm
+        self.require_year = require_year
         self.require_fp = require_fp
 
         self.playlist_type = self.get_playlist_type()
@@ -63,7 +76,15 @@ class PlaylistParser(object):
     def parse(self):
         if self.verbose:  # pragma: no cover
             print("Parsing...")
-        self.tracks = self._parser(self.file_path, require_fp=self.require_fp, verbose=self.verbose)  # type: ignore
+        self.tracks = self._parser(
+            self.file_path,
+            require_title=self.require_title,  # type: ignore
+            require_duration=self.require_duration,  # type: ignore
+            require_bpm=self.require_bpm,  # type: ignore
+            require_year=self.require_year,  # type: ignore
+            require_fp=self.require_fp,  # type: ignore
+            verbose=self.verbose,  # type: ignore
+        )  # type: ignore
         self.is_parsed = True
         if self.verbose:  # pragma: no cover
             print(f"Found {len(self.tracks)} tracks.")