+++ src/playlistparser/parsers/engine.py
@@ -3,7 +3,16 @@ import csv
 from ..track import Track
 
 
-def parser(file_path, *, require_fp=False, verbose=False):
+def parser(
+    file_path,
+    *,
+    require_title=True,
+    require_duration=False,
+    require_year=False,
+    require_bpm=False,
+    require_fp=False,
+    verbose=False,
+):
     """
     Engine supports:
     - title
@@ -14,12 +23,6 @@ def parser(file_path, *, require_fp=False, verbose=False):
     - bpm
     - file_path
     """
-    # Engine v1.6 always includes file paths
-    if require_fp:  # pragma: no cover
-        pass
-    else:  # pragma: no cover
-        pass
-
     with open(file_path) as file:
         reader = csv.DictReader(file)
         tracks = []