+++ CHANGES.md
@@ -2,7 +2,7 @@
 
 ## 3.0.0-beta.7 (unreleased)
 
-- Feat: added ability to require file_path (In #6)
+- Feat: added ability to require title, duration, year, bpm, and file_path (In #6 and #7)
 
 ## 3.0.0-beta.6 (2022-03-25)